fix: broken local build due to sed -i warning (#206)
The workaround to replace sdk versions in
[replace-version.sh](https://github.com/launchdarkly/js-core/blob/main/scripts/replace-version.sh#L14)
does not work locally on OS X because the sed command is not posix
compliant. This breaks yarn build locally on os x.

On os x, `sed -i` expects the next argument to a backup file extension.
On Linux/Unix, it does not.

This ticket implements a workaround for the workaround. As you can see,
this is a rabbit hole which eventually is best solved by not using this
script at all.