feat: edge sdks should send events to bulk/environment endpoint (#256)
Edge sdks use clientSideID not sdk-key. We should be able to post events
to the eventsUri/bulk/clientSideID endpoint for edge sdks without
needing to specify the sdk key in the authorization header.

This is an alternate solution to #217. Both prs are attempting to enable
events for edge sdks.

Don't be alarmed with the filecount, it's mostly shell file fixes
unrelated to this to include a directive which seems to be needed for
local build.

---------

Co-authored-by: LaunchDarklyReleaseBot <launchdarklyreleasebot@launchdarkly.com>