feat: Implement common support for auto environment attributes. (#355)
This PR adds support for auto env attributes for the common library.
This was originally part of #354. That big PR is being broken down into
smaller PRs for the benefit of better changelog through conventional
commits.

Note I have intentionally omitted the `common/jest.config.js` change
because that is part of a larger change to the mocks api. This change
is in #356.