fix: Add LDOptions.application name and versionName. (#358)
Adds support for name and versionName in addition to id and version. I
have also renamed the client sdk config option from `application` to
`applicationInfo`. I have kept the `ApplicationTags` type though,
because that's a shared type with the server sdks and it's a much bigger
change to rename that.