fix: Log errors when there is no error event handler. (#495)
When there are no error event handlers the server SDKs should log the
errors instead.

This behavior was missed during the typescript re-write.

We may want to consider if we always want to log errors, but I
understand the logic of this. If you are listening to this, then you may
also be logging it, and if you are not logging it, then you are at least
getting the event.

I imagine most consumers are not using the error events.