fix: Make it more clear what is happening when an event source is connecting. (#518)
The event source was treating the initial connection as a re-connection.
This results in confusing messaging.

This also changes the retrying event, which we do not expose.

When we convert our polyfill to typescript we should use it instead and
allow passing a connection method. The reason this is being used is
theoretically an issue with fetch not streaming in RN, so it uses an XHR
instead.