feat: Remove event target shim. (#545)
Unlike the EventEmitter in the node SDK the EventTarget was not
providing an interface directly on the LDClient. The LDClient has its
own off/on interface which is not directly implemented via an event
target. Which means removing it simplifies the implementation instead of
complicating it.

Second it is in the common code and node in a leaf implementation. Which
means it requires a polyfill where it is not supported. Like when using
hermes with React Native.

Typically we dispatch using a micro-task, but EventTarget dispatches
synchronously. For now this maintains the synchronous behavior.

Fixes: #412