feat: Correct client evaluation typings. (#554)
Some code was refactored to make server and client share evaluation
results. This should not have been done because the results are not the
same in client and server SDKs. Server SDKs can always produce a reason
and client SDKs cannot.

This meant that the typing said that reason was required in the client
SDK, but it could be null.

This is a 'feat' to ensure a minor version bump in case of minor
incompatibilities.