chore: Apply linting to root of project. (#557)
There is a `check` command at the root, it lints things that are out of
scope of the individual linters. After this baseline we may want to
consider updating some of the per-project linting rules.

This linting is in its own chore PR to prevent creating a release for
these changes.