chore: Scaffold browser SDK project. (#555)
For now I am using Vite to build. If we encounter problems with this we
can move to a rollup based build like we are using in other projects.
Vite is easier to use and is reasonably popular, so I think it is a good
starting point.

Currently all tests in the monorepo use jest, so this continues to use
jest. We may want to consider, in the future, if we want to move to
vitest if it supports everything we need. This would make the testing
configuration simpler.