chore: Build releases in dependency order. (#563)
This contains several conditions that appear to be illogical, but are
required by github actions. For some background this issue can be read:
https://github.com/actions/runner/issues/491

`always()`:

Imagine two packages. Package "a" and package "b".

"b" depends on "a" and we want to automatically release "a" and "b"
using release please. We make "b" `needs` "a", so that "b" will build
after "a". But if we do that, then "b" will not build unless "a" also
builds. We often need to release "b" even though there are no changes to
"a".

By using `if: always()` we can cause "b" to run after "a" even when "a"
doesn't run.

`failure()` and `canceled()`.

With `always()` package "b" will be built and released even if package
"a" failed to build. You would expect that to only require `failure()`,
but it seems that it only works if you have `canceled()` as well. (It
may be we don't actually need failure.)

We want the build order to be like this.
![dep-order-2
drawio](https://github.com/user-attachments/assets/7d499cdb-1163-442b-aefa-7bac01ad560b)