fix: Correctly handle null values in JSON variations. (#569)
During the typescript implementation null values were removed during
JSON de-serialization. This was over-zealous as variations can be JSON
which contains null values.

This retains null value removal for everything aside from variations.

The reason this was originally done is to simplify all code which
interacts with the data model (It only needs to check undefined versus
null/undefined.). It also simplifies the ability to produce a compact
representation that omits any null fields.

In the future we may want to consider removing this behavior.

Fixes #568