feat: Add support for conditional event source capabilities. (#577)
This adds the ability to specify what capabilities an even source
supports.

The idea is, for something like report, the common code can check if the
event source supports it. If it is configured, and the event source
doesn't support it, then it can take appropriate action.

The browser SDK can calculate this at runtime by checking for a polyfill
and filling out the appropriate capabilities.