feat: Adds support for REPORT. (#575)
**Requirements**

- [x] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [x] I have validated my changes against all supported platform
versions

**Related issues**

251674

**Describe the solution you've provided**

- Added useReport to LDOptions.
- Piped through client to `StreamingProcessor` constructor.
- Cloned `StreamingProcessor` and associated tests from common-internal
to sdk-client. Refactored it to support taking in a
`StreamingDataSourceConfig` to narrow provided dependencies instead of
ClientContext which contained many concerns. `StreamingProcessor` now
uses a paths provider to form URI and internally decides where to put
the serialized context (url vs body).
- Refactored LDClientImpl tests to not use MockStreamingProcessor to
reduce global mock usage.
- fix: polling did not include the context in the body.

**Describe alternatives you've considered**

Initially tried making changes in common-internal, but the existence of
the client context in the URL or body for client situations made that
not ideal.

---------

Co-authored-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>