feat: Refactor data source connection handling. (#591)
Most of this PR is refactoring to move data source handling our of the
common code. This also moves functionality, such as `setConnectionMode`
out of the common code.

Currently the MobileDataManager is in the RN SDK, but it should be moved
to some kind of non-browser common when we implement SDKs like node or
electron.

A large amount of the code in this PR is tests for data source handling
and moving existing code.

This PR does not add automatic starting/stopping of the stream.

BEGIN_COMMIT_OVERRIDE
feat: Refactor data source connection handling. 
feat: Add support for js-client-sdk style initialization.
END_COMMIT_OVERRIDE