feat: Apply private property naming standard. Mangle browser private properties. (#620)
1. Update private properties and methods to start with a leading
underscore.
2. Update browser terser config to mangle properties that start with an
underscore that are not `_meta`. (It would be nice to have a more
elegant way to do this, but this is what we have.)
3. Add linting rules for typescript naming conventions.
5. Fix lint issues from rules.

Conventions

Private properties and methods start with _ and CamelCase case names.
Public methods and properties use CamelCase and do not start with an
underscore.
Public static properties use PascalCase and do not start with an
underscore. (This is negotiable, but required fewer changes.)

sdk-764