feat: Add support for inspectors. (#625)
The flag usage and identify were able to be implemented in terms of
hooks.

The flag configuration related inspectors had to be implemented
directly. It would be ideal to figure out how to gauge demand for this
functionality and simply remove it when we deprecated hooks if demand is
low.