feat: Enhance basic logger destination support. (#650)
This adds support to the common basic logger that makes it more suitable
for client SDKs.

Instead of a single destination for all logs you can now control the
destination per log level. This allow you to use all the default
formatting, log level handling, and log tagging while also getting logs
that are nice for use in a browser.