feat: Random uuid for telemetry package. (#689)
Originally this implementation was done for the telemetry package. When
implementing the browser package we needed the same implementation.

Currently this cannot be easily shared, but a task has been created to
followup and change that.

This is now a copy of the version in the browser package.

Subsequent PRs will make use of this functionality.