feat: Add basic logging support for browser-telemetry. (#736)
Adds the ability to specify a logger and to also use the base SDK logger
when possible. The base SDK of a 3.x version does not expose the logger,
but 4.x will.

There is a minimal replication of the warning level of the logging
interface to allow compatibility with both the 3.x SDK and the 4.x SDK.

Prefixing is done at message time instead of as part of the logger.
Potentially it could be ideal to do it in the logger, but this approach
makes it clear that the interpolation supported by most browser loggers
will not come into play. (Where if the logger does this prefixing it is
either complex or you lose the sprintf style formatting support.)

Some earlier log messages has been added, but a logger instance was not
yet available.