feat: Add the ability to filter errors. (#743)
This PR adds support for error level filtering. This is the highest
level filter type capable of filtering on any data.

Other filters are provided to simplify the implementation of filtering
across all error types. Breadcrumb filters could be implemented in terms
of an error filter, but they would have higher complexity and different
performance characteristics. For example with a breadcrumb filter we
filter that breadcrumb regardless of how many events it may appear in,
versus having to filter that same breadcrumb each time an event is
captures.

Custom url filters operate at the HTTP capture level similarly reducing
the frequency of redaction.