fix: Akamai data is fetched for the lifetime of the SDK connection
When the SDK retrieves data from the EdgeKV, it does so using a single
sub-request per call to `variation`, `variationDetail`, or
`allFlagsState`. The problem is that Akamai imposes a limit of 4
sub-requests per handler event.

If a customer evaluates more than 4 distinct flags during the handling
of a single event, subsequent flag lookups would fail. To combat this,
we now cache the flag values the first time an evaluation is requested,
and we keep that cache in memory for the full duration of the EdgeWorker
execution.