fix: Fix a bug where the incorrect src lines may have been captured. (#792)
Resolves a bug where the srcLine and before/after context would have
been populated incorrectly.

We would assume the src line would be in the middle of the source
context, but when a number of lines greater than the context window
don't exist, it will not always be in the middle.

The fix extends TraceKit with a srcStart field that indicates the line
the context window starts on. This can then be used to calculate the
offset to the originating line.