feat: Environment ID support for hooks (#823)
**Requirements**

- [x] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [x] I have validated my changes against all supported platform
versions

**Describe the solution you've provided**

If present, `environmentId` is now passed to the HookRunner in the
evaluation series.

For streaming connections:
- Response headers are now attached to the `open` event
(https://github.com/launchdarkly/js-eventsource/pull/33)
- `StreamingProcessor` passes these headers to the stream listeners via
`processJson`. The listener for the PUT event extracts the
`environmentId` from the headers and passes this as initialization
metadata to the underlying feature store.

For polling connections:
- `PollingProcessor` retrieves the response headers via the underlying
`Requestor` and extracts the `environmentId` from the headers and passes
this as initialization metadata to the underlying feature store.

LDClient will then call `getInitMetaData()` on the underlying feature
store (if the feature store supports it) when executing a hook and pass
`environmentId` in the execution series data.

Currently only `InMemoryFeatureStore` has been modified to support
initialization metadata. This functionality can be added to other
feature stores by modifying `init()` to accept the optional
`initMetadata` parameter and implementing the optional
`getInitMetadata()` method.