fix: Client SDKs should use wrapper information. (#836)
Wrappers are accessed through the platform, but react-native and the
browser implementation were not propagating these to the platform.

The first implementation was the node platform, which handled this
correctly. The Cloudlare and RN platform didn't, and then the RN
platform was used as reference for implementing the browser platform.

So, this also affects edge and a similar update should be considered for
those SDKs.