feat: Add client-side SDK plugin support. (#834)
Adds plugin support for react native and browser.

Types are common, as well as most functionality, but SDK specific
implementation is required. This is because the leaf-node SDKs have
different types for the LDClient. Also the SDK needs control over the
ordering of operations that happen during construction in order to
register plugins at the correct point.

There is a bit more complexity in this implementation than would be
required for most SDKs. Some refactoring of the base client
implementation, potentially to use composition instead of inheritance,
may allow for some better sharing here in the long-term.

Initially I tried using generics in the base implementation, to account
for the different client implementations, but this introduced a lot of
changes to all the layers and didn't play well with configuration, which
also then needs to be generic.

The approach I took was to have the configuration in the individual SDKs
and then use generics for utility methods to collect hooks and register
plugins. In the long-term this would also benefit from the base
implementation using generics for the configuration, which would allow a
shared configuration and validation.