chore: Add implementation for async task queue. (#841)
This PR adds an async task queue. This queue is roughly based on the
flutter implementation, but with support for non-shedable tasks like
iOS.

https://github.com/launchdarkly/flutter-client-sdk/blob/01841d65d8eb9dbf7cdc1c29b76f2ac8a699fb05/packages/common/lib/src/async/async_single_queue.dart

https://github.com/launchdarkly/ios-client-sdk/blob/082430ff008d76d503327c52b27e3b2af99c3ef2/LaunchDarkly/LaunchDarkly/ServiceObjects/SheddingQueue.swift

This is a `chore`, because the type is unused and doesn't need a
release. A subsequent PR will use it for functionality.

This queue is going to be used for an implementation of the client-side
identify spec.

https://github.com/launchdarkly/sdk-specs/tree/main/specs/CSI-client-side-identify

(Spec is private.)

Available for review by anyone, I added Matthew for reference to the iOS
SDK in case he was interested.