feat: Add plugin support for node. (#880)
This adds plugin support for the node server SDK.

Plugin support is leaf-SDK specific because plugins use the interface of
the SDK. The interface of the SDK has slightly different capabilities
based on implementation.

Unfortunately this isn't a simple problem to tackle with regards to
organization. We could use composition with the common functionality
being generic over the SDK interface. Doing so would require some
breaking changes to all the SDK implementation, but could be worth it in
the future.

For not though we have to settle for some complexity with the plugin
interface being defined per-sdk and the call to the initialization code
being per-sdk.

This was able to re-use the generic types and functions which means this
isn't very much code.

This SDK had another complexity which is the event emitter. For
convenient implementation of events we were using a mixin, but this
mixin approach means we didn't have the full interface of the SDK
available at the construction time of the SDK. So this PR moves the
emitter implementation into the node client.