feat: additional FDv2 functionality (#898)
This PR is nearly an empty commit and removes a completed TODO with a
`feat` commit to trigger a release.

Up to this point, no FDv2 changes in shared/common had been made in a
`feat` commit, so`js-sdk-common` didn't have a release triggered since
May 21st. This was overlooked as additional functionality was merged
into main in other packages that did get released.